updating readme for final submission
renamed sections, separating out my bibliography from my actual
references that were used in the technology review. i also added
an "abstract" to the readme which is really just the introduction
from the technology review itself. i also added myself as an author
and revised the topic description slightly to more accurately
represent the final subject of study. finally, i added a link to
the final pdf in the readme.